Separated out startWindow code
